refactor: extract group management UI into GroupViews.swift
Move the group management surface (GroupDetailsSheet, GroupMemberRow,
GroupDiagnosticsValueRow, the group-image picker/results, and the
GroupImageSearchResult.previewURL helper) out of MessengerShellView.swift
into GroupViews.swift. Top-level declarations promoted to internal; the
previewURL extension stays file-private since its only consumer moved with
it.

Also make conversationHeaderChatInfoButtonOpensGroupDetailsSheet
location-robust: it previously bounded the ConversationHeader snippet using
"private struct GroupDetailsSheet" as an end-marker in MessengerShellView,
which the extraction removed. It now finds ConversationHeader in whichever
Views file holds it and bounds the snippet by the next declaration, so the
guard survives this and the upcoming composer extraction. The user-facing
invariant it checks (info button wired to the group details sheet) is
unchanged.

Build green; test suite unchanged (same two pre-existing image-downsampling
failures, unrelated to view extraction).
